+++ app/scripts/commons/components/messageNoLanguage.js
@@ -1,5 +1,5 @@
 'use strict'
 
 angular.module('serinaApp').component('messageNoLanguage', {
-  templateUrl: 'views/commons/directives/message-no-language.html'
+  templateUrl: 'views/commons/components/message-no-language.html'
 })