+++ app/scripts/commons/components/sideMenu.js
@@ -1,5 +1,5 @@
 'use strict'
 
 angular.module('serinaApp').component('sideMenu', {
-  templateUrl: 'views/commons/directives/side-menu.html'
+  templateUrl: 'views/commons/components/side-menu.html'
 })