+++ app/scripts/commons/components/spinner.js
@@ -4,5 +4,5 @@ angular.module('serinaApp').component('spinner', {
   bindings: {
     loading: '<'
   },
-  templateUrl: 'views/commons/directives/spinner.html'
+  templateUrl: 'views/commons/components/spinner.html'
 })